Refactor symbol deduplication logic
Previously symbols were only deduplicated if the original symbol
was of the same type. For example if a function symbol was created
from parsing the .mdebug symbol table, if there was also a .sndata
symbol table, a redundant label symbol would be created at the
same position.

This was half fixed by the last commit that allowed the SNDLL
parser to guess the type of symbol, but the existing symbol
deduplication was also somewhat messy, so this cleans it up.